+++ library_common/src/main/java/com/firewood/context/BaseContext.java
@@ -2,9 +2,9 @@ package com.firewood.context;
 
 public class BaseContext {
 
-    public static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
-    public static void setCurrentId(Long id){threadLocal.set(id);}
-    public static Long getCurrentId(){return threadLocal.get();}
+    public static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
+    public static void setCurrentId(Integer id){threadLocal.set(id);}
+    public static Integer getCurrentId(){return threadLocal.get();}
     public static void removeCurrentId(){threadLocal.remove();}
 
 }
